feat : 멤버쉽 업그레이드 기능 추가

+++ src/main/java/store/myproject/onlineshop/domain/membership/MemberShip.java
@@ -30,7 +30,7 @@ public class MemberShip extends BaseEntity {
 
     private Float discountRate;
 
-    private Long baseline;
+    private BigDecimal baseline;
 
     public void updateMemberShip(MemberShipUpdateRequest updateRequest) {
         this.level = updateRequest.getLevel();
