+++ src/main/java/store/myproject/onlineshop/domain/membership/dto/MemberShipDto.java
@@ -3,6 +3,8 @@ package store.myproject.onlineshop.domain.membership.dto;
 import lombok.*;
 import store.myproject.onlineshop.domain.customer.Level;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
@@ -11,7 +13,7 @@ public class MemberShipDto {
 
     private Level level;
 
-    private Long baseline;
+    private BigDecimal baseline;
 
     private Float discountRate;
